requirements.txt
cryptography==40.0.1
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pytest==7.3.0
requests==2.28.2